release version 3.0.0-alpha3

+++ Installation/release.sh
@@ -3,6 +3,24 @@ set -e
 
 TAG=1
 BOOK=1
+BUILD=1
+SWAGGER=1
+EXAMPLES=1
+
+if [ "$1" == "--no-build" ];  then
+  BUILD=0
+  shift
+fi
+
+if [ "$1" == "--no-swagger" ];  then
+  SWAGGER=0
+  shift
+fi
+
+if [ "$1" == "--no-examples" ];  then
+  EXAMPLES=0
+  shift
+fi
 
 if [ "$1" == "--no-commit" ];  then
   TAG=0
@@ -51,14 +69,16 @@ if [ `uname` == "Darwin" ];  then
   CMAKE_CONFIGURE="${CMAKE_CONFIGURE} -DOPENSSL_ROOT_DIR=/usr/local/opt/openssl -DCMAKE_OSX_DEPLOYMENT_TARGET=10.11"
 fi
 
-echo "COMPILING"
-rm -rf build && mkdir build
+if [ "$BUILD" == "1" ];  then
+  echo "COMPILING"
+  rm -rf build && mkdir build
 
-(
-  cd build
-  cmake .. ${CMAKE_CONFIGURE}
-  make -j 8
-)
+  (
+    cd build
+    cmake .. ${CMAKE_CONFIGURE}
+    make -j 8
+  )
+fi
 
 echo "LINTING"
 ./utils/jslint.sh
@@ -75,11 +95,15 @@ git add -f \
   js/common/bootstrap/errors.js \
   CMakeLists.txt
 
-echo "SWAGGER"
-./utils/generateSwagger.sh
+if [ "$SWAGGER" == "1" ];  then
+  echo "SWAGGER"
+  ./utils/generateSwagger.sh
+fi
 
-echo "EXAMPLES"
-./utils/generateExamples.sh
+if [ "$EXAMPLES" == "1" ];  then
+  echo "EXAMPLES"
+  ./utils/generateExamples.sh
+fi
 
 echo "GRUNT"
 (
